+++ cg/meta/demultiplex/demux_post_processing.py
@@ -1,11 +1,13 @@
 import logging
 from pathlib import Path
+from typing import List, Iterable
 
 import alchy
 from cg.apps.cgstats.crud import create, find
 from cg.apps.cgstats.stats import StatsAPI
 from cg.constants.cgstats import STATS_HEADER
 from cg.meta.demultiplex import files
+from cg.models.cgstats.stats_sample import StatsSample
 from cg.models.demultiplex.demux_results import DemuxResults
 
 LOG = logging.getLogger(__name__)
@@ -35,8 +37,33 @@ class DemuxPostProcessingAPI:
         create.create_novaseq_flowcell(manager=self.stats_api, demux_results=demux_results)
 
     @staticmethod
-    def fetch_report_data(flowcell_id: str, project_name: str) -> alchy.Query:
+    def fetch_report_data(flowcell_id: str, project_name: str) -> List[StatsSample]:
         return find.project_sample_stats(flowcell=flowcell_id, project_name=project_name)
+
+    @staticmethod
+    def sample_to_report_line(stats_sample: StatsSample, flowcell_id: str) -> str:
+        lanes: str = ",".join(str(lane) for lane in stats_sample.lanes)
+        return "\t".join(
+            [
+                stats_sample.sample_name,
+                flowcell_id,
+                lanes,
+                ",".join(str(unaligned.read_count) for unaligned in stats_sample.unaligned),
+                str(stats_sample.read_count_sum),
+                ",".join(str(unaligned.yield_mb) for unaligned in stats_sample.unaligned),
+                str(stats_sample.sum_yield),
+                ",".join(str(unaligned.q30_bases_pct) for unaligned in stats_sample.unaligned),
+                ",".join(str(unaligned.mean_quality_score) for unaligned in stats_sample.unaligned),
+            ]
+        )
+
+    @staticmethod
+    def get_report_lines(stats_samples: List[StatsSample], flowcell_id: str) -> Iterable[str]:
+        """Convert stats samples to format lines ready to print"""
+        for stats_sample in sorted(stats_samples, key=lambda x: x.sample_name):
+            yield DemuxPostProcessingAPI.sample_to_report_line(
+                stats_sample=stats_sample, flowcell_id=flowcell_id
+            )
 
     @staticmethod
     def write_report(report_path: Path, report_content: alchy.Query):
@@ -61,6 +88,8 @@ class DemuxPostProcessingAPI:
                     )
                     + "\n"
                 )
+
+    def get_report_data(self, demux_results: DemuxResults) -> Iterable[str]:
 
     def create_cgstats_reports(self, demux_results: DemuxResults) -> None:
         """Create a report for every project that was demultiplexed"""