+++ CHANGELOG.md
@@ -13,6 +13,17 @@ Try to use the following format:
 ### Changed
 ### Fixed
 
+## [21.4.1]
+### Changed
+- Change how samples are fetched for cgstats select command
+### Fixed
+- Bug when fetching info with `cg demultiplex select`-command
+
+
+## [21.4.0]
+### Added
+- A column in customer-group admin view in to show customers in the group
+
 ## [21.3.1]
 ### Fixed
 - PDC backup destination server for 2500 flowcells